refactor(core-types): Layout strides switch from usize to isize (StrideVec)
Foundation for negative-stride view ops (Op::Flip-as-view next). New
`StrideVec = SmallVec<[isize; 6]>` type alias; Layout::stride field +
return type follow. New `Layout::stride_unsigned()` debug-asserts
non-negative and returns DimVec for kernels that don't yet handle
negative strides — most of them, since auto-Contiguize ensures
positive strides on kernel input today.

StridedIndex stores `next_storage_index: Option<isize>` and yields
`usize` at the iterator boundary, so it works uniformly for positive
and negative strides.

Tensor::stride() return type changes &[usize] → &[isize] (one breaking
public API signature; users who don't care about Flip can stride
through stride_unsigned() at the call site or cast at boundaries).

Three helper fns added in fuel-cuda-backend/src/storage.rs to
concentrate the cast at the FFI boundary:
- dims_strides_usize(layout)
- dims_strides_strides_usize(dims, lhs, rhs)
- conv_dims_strides_usize(dims, inp, k)

Tests: fuel-core-types 14 ✓ • fuel-cpu-backend 145 ✓ • fuel-graph 147
✓ • fuel-core 135 ✓ • fuel-reference-backend + fuel-graph-cpu 184 ✓.
Workspace `cargo check` clean except for pre-existing unrelated
breakage (fuel-flash-attn-cuda storage_and_layout debt, fuel-onnx,
fuel-metal-* on Windows, fuel-flash-attn-v3-cuda-sys sm90).

+++ fuel-core-types/src/strided_index.rs
@@ -3,23 +3,31 @@ use smallvec::smallvec;
 
 /// An iterator over offset position for items of an N-dimensional arrays stored in a
 /// flat buffer using some potential strides.
+///
+/// `next_storage_index` is held as `isize` internally so negative
+/// strides (e.g. from `Op::Flip`) work uniformly: each step adds
+/// `stride[k]` (signed) to the current offset, and the iteration
+/// invariant is that the cumulative offset stays in
+/// `[0, storage.len())` — a property the producing Layout's
+/// `start_offset` guarantees by construction. Yielded indices are
+/// cast to `usize` at the boundary.
 #[derive(Debug)]
 pub struct StridedIndex<'a> {
-    next_storage_index: Option<usize>,
+    next_storage_index: Option<isize>,
     multi_index: DimVec,
     dims: &'a [usize],
-    stride: &'a [usize],
+    stride: &'a [isize],
     remaining: usize,
 }
 
 impl<'a> StridedIndex<'a> {
-    pub fn new(dims: &'a [usize], stride: &'a [usize], start_offset: usize) -> Self {
+    pub fn new(dims: &'a [usize], stride: &'a [isize], start_offset: usize) -> Self {
         let elem_count: usize = dims.iter().product();
         let next_storage_index = if elem_count == 0 {
             None
         } else {
             // This applies to the scalar case.
-            Some(start_offset)
+            Some(start_offset as isize)
         };
         StridedIndex {
             next_storage_index,
@@ -57,7 +65,7 @@ impl Iterator for StridedIndex<'_> {
                 next_storage_index += stride_i;
                 break;
             } else {
-                next_storage_index -= *multi_i * stride_i;
+                next_storage_index -= (*multi_i as isize) * stride_i;
                 *multi_i = 0
             }
         }
@@ -67,7 +75,9 @@ impl Iterator for StridedIndex<'_> {
         } else {
             None
         };
-        Some(storage_index)
+        // Cast to usize at the boundary; the Layout's construction
+        // invariant guarantees the offset is non-negative.
+        Some(storage_index as usize)
     }
 
     #[inline]
